platform/x86: Update SoCWatch driver for 5.9 pull
Update to SoCWatch driver version 2.13 which has a fix
to the PMT driver interface for the 5.9 pull.
PMT data is not 16 bytes. The first 8 bytes are metadata used
to identify the source; It contains the GUID of the PMT endpoint,
the unique runtime endpoint ID and the PCI slot. The second 8 bytes
is the actual read payload data

 Signed-off-by: Faycal Benmlih <faycal.benmlih@intel.com>